fix: :bug: changed trial balance system
previously there was two ways to get the trial balance, one was default trial balance which displayed the trial balance as on the current date and the other asked for the as on date. Now changed this thing to always ask for as on date, and empty input will be taken as current date.