feat: :sparkles: added date suggestion or default input feature
the user is asked to enter the 'to' and 'from' date or 'as on' date for viewing statements. Now, added a default input feature. It shows the today's date for the 'to' and 'as on' date input and show the date of 1 week ago for the 'from' date. If the user enters an empty input, the defaults will be selected automatically.